Defragmentation in a btree page
This corresponds to defragmentPage() of SQLite. This moves whole content
to a newly allocated temporary buffer and replace old buffer with the
new buffer. This copies exactly a page size of memory.

The original defragmentPage() copies whole cell content area to a
temporary memory space and move contents back to the original buffer.
This cause less memory copy than the prsqlite way if the size of the
cell content area is less than half of page. However less likely.

The prsqlite way is possible on prsqlite because it is guaranteed that
there is no reference (including pointer) to the updating whole page
(including the database header in page 1).