Split a table leaf page into 3 pages
There can be a case when a new cell can't fit in both splitted pages. In
that case, we need to split the page into 3 pages.

This only happens for a table leaf page because table leaf cells can be
more than half of page size while table interior cells is less than or
equal to 13 bytes and index cells has at most about quarter of usable
size.